[GN-128] 404페이지 재연결 및 코드 수정
