+++ src/app/(DefaultLayout)/(MainLayout)/activities/[id]/components/ActivityInfo.tsx
