+++ src/app/(DefaultLayout)/(MainLayout)/activities/[id]/components/ActivitySkeleton.tsx
