+++ src/app/(DefaultLayout)/(MainLayout)/activities/[id]/components/Description.tsx
