+++ src/app/(DefaultLayout)/(MainLayout)/activities/[id]/components/ImageGallery.tsx
