+++ src/app/(DefaultLayout)/(MainLayout)/activities/[id]/components/KakaoMap.tsx
