+++ src/app/(DefaultLayout)/(MainLayout)/activities/[id]/components/reservation/ReservationMobile.tsx
