+++ src/app/(DefaultLayout)/(MainLayout)/activities/[id]/components/reservation/ReservationMobileCnt.tsx
