+++ src/app/(DefaultLayout)/(MainLayout)/activities/[id]/components/reservation/SlidePanel.tsx
